Anomaly #47860: Pricing little refactor, created service a move anything related to pricing in it. Changed private method to protected, change public method that are not meant to be used externally of the service to protected. Re-do the origin fix of #46204, translations
