--- axelor-base/src/main/java/com/axelor/apps/base/service/PricingService.java
@@ -1,34 +1,0 @@
-/*
- * Axelor Business Solutions
- *
- * Copyright (C) 2021 Axelor (<http://axelor.com>).
- *
- * This program is free software: you can redistribute it and/or  modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.axelor.apps.base.service;
-
-import com.axelor.apps.base.db.Company;
-import com.axelor.apps.base.db.Pricing;
-import com.axelor.apps.base.db.Product;
-import com.axelor.apps.base.db.ProductCategory;
-import com.axelor.db.Query;
-
-public interface PricingService {
-
-  public Query<Pricing> getPricing(
-      Product product,
-      ProductCategory productCategory,
-      Company company,
-      String modelName,
-      Pricing parentPricing);
-}
