+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/pricing/PricingService.java
@@ -1,0 +1,42 @@
+package com.axelor.apps.sale.service.pricing;
+
+import com.axelor.apps.base.db.Company;
+import com.axelor.apps.base.db.Pricing;
+import com.axelor.apps.base.db.PricingLine;
+import com.axelor.apps.base.db.Product;
+import com.axelor.apps.base.db.ProductCategory;
+import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.sale.db.SaleOrderLine;
+import com.axelor.db.Query;
+import com.axelor.exception.AxelorException;
+import java.util.Optional;
+
+public interface PricingService {
+
+  void computePricingScale(SaleOrder saleOrder, SaleOrderLine orderLine) throws AxelorException;
+
+  Query<Pricing> getPricing(
+      Product product,
+      ProductCategory productCategory,
+      Company company,
+      String modelName,
+      Pricing parentPricing);
+
+  /**
+   * Method that get the default pricing of saleOrderLine.
+   *
+   * @return true if it has, else false
+   */
+  Optional<Pricing> getDefaultPricing(SaleOrder saleOrder, SaleOrderLine saleOrderLine);
+
+  /**
+   * Methods that get the pricingLine that matches with in pricing.
+   *
+   * @param saleOrder
+   * @param saleOrderLine
+   * @param pricing
+   * @throws AxelorException
+   */
+  Optional<PricingLine> getPricingLine(
+      SaleOrder saleOrder, SaleOrderLine saleOrderLine, Pricing pricing) throws AxelorException;
+}
