+++ axelor-sale/src/main/resources/i18n/messages_en.csv
@@ -613,6 +613,7 @@
 "Year 1 To Date",,,
 "Year 2 To Date",,,
 "Yearly",,,
+"You are using a product for which the '%s' pricing should be applied.</br>However, it could not be applied.</br>Please check your pricing if this does not seem normal.",,,
 "You have to choose at least one sale order",,,
 "You have to choose at least one sale quotation",,,
 "You must configure a script to fill the created product code.",,,
