+++ axelor-sale/src/main/resources/i18n/messages_fr.csv
@@ -613,6 +613,7 @@
 "Year 1 To Date","Année 1 à date",,
 "Year 2 To Date","Année 2 à date",,
 "Yearly","Annuel",,
+"You are using a product for which the '%s' pricing should be applied.</br>However, it could not be applied.</br>Please check your pricing if this does not seem normal.","Vous utilisez un produit pour lequel le barème '%s' devrait être appliqué.</br>Cependant, celui-ci n'a pas pu être appliqué.</br>Merci de vérifier votre barème si cela ne vous semble pas normal.",,
 "You have to choose at least one sale order",,,
 "You have to choose at least one sale quotation",,,
 "You must configure a script to fill the created product code.","Veuillez configurer un script pour renseigner le code du produit créé.",,
